Move build_subscription_payload to envs model
