+++ atat/routes/applications/settings.py
@@ -6,7 +6,6 @@ from flask import (
     request as http_request,
     url_for,
 )
-from secrets import token_urlsafe
 
 from .blueprint import applications_bp
 from atat.domain.exceptions import AlreadyExistsError
@@ -15,7 +14,6 @@ from atat.domain.applications import Applications
 from atat.domain.application_roles import ApplicationRoles
 from atat.domain.audit_log import AuditLog
 from atat.domain.csp.cloud.exceptions import GeneralCSPException
-from atat.domain.csp.cloud.models import SubscriptionCreationCSPPayload
 from atat.domain.common import Paginator
 from atat.domain.environment_roles import EnvironmentRoles
 from atat.domain.invitations import ApplicationInvitations
@@ -532,27 +530,6 @@ def resend_invite(application_id, application_role_id):
     )
 
 
-def build_subscription_payload(environment) -> SubscriptionCreationCSPPayload:
-    csp_data = environment.portfolio.csp_data
-    parent_group_id = environment.cloud_id
-    invoice_section_name = csp_data["billing_profile_properties"]["invoice_sections"][
-        0
-    ]["invoice_section_name"]
-
-    display_name = (
-        f"{environment.application.name}-{environment.name}-{token_urlsafe(6)}"
-    )
-
-    return SubscriptionCreationCSPPayload(
-        tenant_id=csp_data.get("tenant_id"),
-        display_name=display_name,
-        parent_group_id=parent_group_id,
-        billing_account_name=app.config["AZURE_BILLING_ACCOUNT_NAME"],
-        billing_profile_name=csp_data.get("billing_profile_name"),
-        invoice_section_name=invoice_section_name,
-    )
-
-
 @applications_bp.route(
     "/environments/<environment_id>/add_subscription", methods=["POST"]
 )
@@ -561,7 +538,9 @@ def create_subscription(environment_id):
     environment = Environments.get(environment_id)
 
     try:
-        payload = build_subscription_payload(environment)
+        payload = environment.build_subscription_payload(
+            app.config["AZURE_BILLING_ACCOUNT_NAME"]
+        )
         app.csp.cloud.create_subscription(payload)
         flash("environment_subscription_success", name=environment.displayname)
