+++ tests/routes/applications/test_settings.py
@@ -12,6 +12,7 @@ from werkzeug.datastructures import ImmutableMultiDict
 from atat.database import db
 from atat.domain.application_roles import ApplicationRoles
 from atat.domain.applications import Applications
+from atat.domain.environments import Environments
 from atat.domain.common import Paginator
 from atat.domain.csp.cloud.azure_cloud_provider import AzureCloudProvider
 from atat.domain.csp.cloud.exceptions import GeneralCSPException
@@ -29,7 +30,6 @@ from atat.models.application_role import Status as ApplicationRoleStatus
 from atat.models.environment_role import CSPRole, EnvironmentRole
 from atat.models.permissions import Permissions
 from atat.routes.applications.settings import (
-    build_subscription_payload,
     filter_env_roles_data,
     filter_env_roles_form_data,
     get_environments_obj_for_app,
@@ -849,66 +849,3 @@ def test_create_subscription_failure(client, user_session, monkeypatch):
     )
 
     assert response.status_code == 400
-
-
-class TestBuildSubscriptionPayload:
-    def test_unique_display_name(self):
-        # Create 2 Applications with the same name that both have an environment named 'Environment'
-        app_1 = ApplicationFactory.create(
-            name="Application", environments=[{"name": "Environment", "cloud_id": 123}]
-        )
-        env_1 = app_1.environments[0]
-        app_2 = ApplicationFactory.create(
-            name="Application", environments=[{"name": "Environment", "cloud_id": 456}]
-        )
-        env_2 = app_2.environments[0]
-        env_1.portfolio.csp_data = {
-            "billing_account_name": "xxxx-xxxx-xxx-xxx",
-            "billing_profile_name": "xxxxxxxxxxx:xxxxxxxxxxxxx_xxxxxx",
-            "tenant_id": "xxxxxxxxxxx-xxxxxxxxxx-xxxxxxx-xxxxx",
-            "billing_profile_properties": {
-                "invoice_sections": [{"invoice_section_name": "xxxx-xxxx-xxx-xxx"}]
-            },
-        }
-        env_2.portfolio.csp_data = {
-            "billing_account_name": "xxxx-xxxx-xxx-xxx",
-            "billing_profile_name": "xxxxxxxxxxx:xxxxxxxxxxxxx_xxxxxx",
-            "tenant_id": "xxxxxxxxxxx-xxxxxxxxxx-xxxxxxx-xxxxx",
-            "billing_profile_properties": {
-                "invoice_sections": [{"invoice_section_name": "xxxx-xxxx-xxx-xxx"}]
-            },
-        }
-
-        # Create subscription payload for each environment
-        payload_1 = build_subscription_payload(env_1)
-        payload_2 = build_subscription_payload(env_2)
-
-        assert payload_1.display_name != payload_2.display_name
-
-    def test_populates_payload_correctly(self, app):
-        application = ApplicationFactory.create(
-            name="Application", environments=[{"name": "Environment", "cloud_id": 123}]
-        )
-        environment = application.environments[0]
-        account_name = app.config["AZURE_BILLING_ACCOUNT_NAME"]
-        profile_name = "fake-profile-name"
-        tenant_id = "123"
-        section_name = "fake-section-name"
-
-        environment.portfolio.csp_data = {
-            "billing_profile_name": profile_name,
-            "tenant_id": tenant_id,
-            "billing_profile_properties": {
-                "invoice_sections": [{"invoice_section_name": section_name}]
-            },
-        }
-        payload = build_subscription_payload(environment)
-        assert type(payload) == SubscriptionCreationCSPPayload
-        # Check all key/value pairs except for display_name because of the appended random string
-        assert {
-            "tenant_id": tenant_id,
-            "parent_group_id": environment.cloud_id,
-            "billing_account_name": account_name,
-            "billing_profile_name": profile_name,
-            "invoice_section_name": section_name,
-        }.items() <= payload.dict().items()
